2009-07-22  Tatsuhiro Tsujikawa  <t-tujikawa@users.sourceforge.net>
	Fixed compiler warning on 32bit systems.
	* src/DownloadCommand.cc

+++ src/DownloadCommand.cc
@@ -115,7 +115,8 @@ bool DownloadCommand::executeInternal() {
 
   size_t bufSize;
   if(segment->getLength() > 0) {
-    if(segment->getPosition()+segment->getLength() <= static_cast<uint64_t>(_fileEntry->getLastOffset())) {
+    if(static_cast<uint64_t>(segment->getPosition()+segment->getLength()) <=
+       static_cast<uint64_t>(_fileEntry->getLastOffset())) {
       bufSize = std::min(segment->getLength()-segment->getWrittenLength(),
 			 BUFSIZE);
     } else {
